refactor: Rename `…e/compile/events/api-gateway/lib/requestValidator.js`
