+++ lib/plugins/aws/package/compile/events/api-gateway/lib/request-validator.js
